pyproject.toml
[project]
name = "alpha-agents"
version = "0.1.0"
description = "Autonomous crypto trading system with multi-agent + XGBoost"
requires-python = ">=3.11"

dependencies = [
    # 에이전트 오케스트레이션
    "langgraph>=0.3.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",

    # 전략 모델
    "xgboost>=2.0.0",
    "lightgbm>=4.3.0",
    "scikit-learn>=1.4.0",

    # 데이터 수집
    "python-binance>=1.0.19",
    "httpx>=0.27.0",

    # 기술적 지표 (brew install ta-lib 선행 필요)
    "TA-Lib>=0.4.28",

    # 데이터 처리
    "pandas>=2.2.0",
    "numpy>=1.26.0",

    # 인프라
    "asyncpg>=0.29.0",
    "redis>=5.0.0",

    # 설정 관리
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",

    # 모니터링/로깅
    "structlog>=24.0.0",
    "prometheus-client>=0.20.0",

    # Phase 2+: 감성 분석 (주석 해제 시 활성화)
    # "sentence-transformers>=3.0.0",
    # "transformers>=4.40.0",
    # "torch>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "config", "storage", "graph", "backtest", "monitoring"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
